Other minor code fixes.

+++ branches/wezzy/linbox/algorithms/opencl-domain-factory.h
@@ -377,7 +377,7 @@ namespace LinBox{
 
 			//Find the max score and calculate a lower bound for device selection
 			long maxScore = *(std::max_element(rankings.begin(), rankings.end()));
-			long lowerBound = maxScore * 0.90;
+			long lowerBound = (long)(maxScore * 0.90);
 
 			//Build OpenCL compute environments only for devices within 10% of the top device.
 			//This should keep run times for identical computations relatively similar across different
